+++ src/pix.h
@@ -26,10 +26,7 @@ struct RGBC_Quad {
   size_t count;
 };
 typedef struct RGBC_Quad RGBC_QUAD;
-// struct Pix32 {
-//	RGBA_QUAD* pixs;
-//	int n;
-//};
+
 struct Pix {
   void *pixs;
   int n;
@@ -55,7 +52,7 @@ extern PIXCMAP *pix_cmap_generate_from_median_cuts(HEAP *heap, int *histo,
                                                    int sigbits);
 extern PIXCMAP *pix_cmap_create(int depth);
 extern int pix_cmap_add_color(PIXCMAP *cmap, int r_val, int g_val, int b_val,
-                              int count);
+                              size_t count);
 #ifdef __cplusplus
 }
 #endif
